Rename files to use underscores in place of dashes
