+++ src/main/angular/app/project_select/project_select.component.html
