+++ src/main/angular/app/services/data.service.provider.ts
@@ -1,6 +1,6 @@
 import {USE_TEST_DATA} from '../../constants';
-import {FakeDataService} from './fake-services/fake-data.service';
-import {HttpService} from './real-services/http.service';
+import {FakeDataService} from './fake_services/fake_data.service';
+import {HttpService} from './real_services/http.service';
 import {HttpClient} from '@angular/common/http';
 import {DataService} from './data.service';
