+++ src/main/angular/app/services/date_utilities.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {ProjectGraphData} from '../../model/project-graph-data';
+import {ProjectGraphData} from '../../model/project_graph_data';
 
 /** Contains some basic utility methods for date wrangling. */
 @Injectable()
