+++ src/main/angular/app/services/real_services/http.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {ProjectGraphData} from '../../../model/project-graph-data';
+import {ProjectGraphData} from '../../../model/project_graph_data';
 import {Project} from '../../../model/project';
 import {DataService} from '../data.service';
