+++ src/main/angular/model/project_graph_data.ts
