+++ src/main/angular/model/recommender_type.ts
